Cache repo results.
This cuts down significantly on the total number of API calls we need to make,
and makes the page response faster if it's been requested recently.

This also progresses towards making it possible to do an analysis of prolific
users, since we can make several requests and cache a little bit further each
time until we can serve the whole thing. :)

See #1 (but it doesn't solve it!).